[android] Simplify Feature conversion
