+++ platform/android/src/style/sources/custom_geometry_source.cpp
@@ -9,7 +9,7 @@
 // C++ -> Java conversion
 #include "../../conversion/conversion.hpp"
 #include "../../conversion/collection.hpp"
-#include "../../geojson/conversion/feature.hpp"
+#include "../../geojson/feature.hpp"
 #include <mbgl/style/conversion/custom_geometry_source_options.hpp>
 
 #include <string>
@@ -155,7 +155,7 @@ namespace android {
             features = rendererFrontend->querySourceFeatures(source.getID(),
                 { {},  toFilter(env, jni::SeizeLocal(env, std::move(jfilter))) });
         }
-        return *convert<jni::Array<jni::Object<Feature>>, std::vector<mbgl::Feature>>(env, features);
+        return Feature::convert(env, features);
     }
 
     jni::Object<Source> CustomGeometrySource::createJavaPeer(jni::JNIEnv& env) {
